+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.0.2] - 2026-05-18
 
 - Make sure ractorized objects are shareable. This requires them to be unusable after closing (or joining) them.
+- Make sure methods don't collide with core methods: #join/#close -> #__join__/#__close__
 
 ## [0.0.1] - 2026-04-14